docker/backend/app/requirements.txt
Flask==2.1.0
Werkzeug==2.1.0
torch==1.13.1
openai-whisper==20230314
ffmpeg-python==0.2.0
numpy==1.24.4